+++ docs/changelog.rst
@@ -2,6 +2,10 @@
 Changelog
 =========
 
+* Added field classes :class:`~django_mysql.models.TinyIntegerField` and
+  :class:`~django_mysql.models.PositiveTinyIntegerField` that use MySQL’s one-byte
+  ``TINYINT`` data type.
+
 4.16.0 (2025-02-06)
 -------------------